lib: move settings to agent/lib
Canonical home moves from scripts/lib to agent/lib. Its two remaining
scripts/lib consumers (i18n.mjs, menu/lang.mjs) pick up the #lib/
alias for now; content is unchanged.
